+ removed rendundant int cast + fixed mouse flickering at a specific x,y + tooltip optimization (now it appears when needed)

+++ Game/Gumps/Controls/Button.cs
@@ -144,7 +144,7 @@ namespace ClassicUO.Game.Gumps.Controls
             SpriteTexture texture = GetTextureByState();
 
 
-            spriteBatch.Draw2D(texture, new Rectangle((int) position.X, (int) position.Y, Width, Height), IsTransparent ? RenderExtentions.GetHueVector(0, false, 0.5f, false) : Vector3.Zero);
+            spriteBatch.Draw2D(texture, new Rectangle(position.X, position.Y, Width, Height), IsTransparent ? RenderExtentions.GetHueVector(0, false, 0.5f, false) : Vector3.Zero);
 
             //Draw1(spriteBatch, texture, new Rectangle((int) position.X, (int) position.Y, Width, Height), -1, 0, IsTransparent ? RenderExtentions.GetHueVector(0, false, 0.5f, false) : Vector3.Zero);
