+++ Game/Gumps/UIGumps/ChatControl.cs
@@ -263,13 +263,13 @@ namespace ClassicUO.Game.Gumps.UIGumps
 
         public override bool Draw(SpriteBatchUI spriteBatch, Point position, Vector3? hue = null)
         {
-            int y = _textBox.Y + (int) position.Y - 6;
+            int y = _textBox.Y + position.Y - 6;
 
             for (int i = _textEntries.Count - 1; i >= 0; i--)
             {
                 y -= _textEntries[i].TextHeight;
 
-                if (y >= (int) position.Y)
+                if (y >= position.Y)
                     _textEntries[i].Draw(spriteBatch, new Point(position.X + 2, y));
             }
